Moved the class for processing and storing title.basics.tsv to a new file. Finally fixed the issue with not filtering adult titles - seems casting to integer doesn't work very well.
